feat: ApplicantDetailPage에 사용되는 컴포넌트들 (ApplicantDetailContent, Header, DataController) (#12)
